+++ demucs/metal/group_norm_gelu.metal
@@ -46,8 +46,9 @@ kernel void group_norm_g1_gelu(
     threadgroup float shared_sqsum[MAX_TGS];
 
     const uint total = C * N;
-    device const SCALAR_T* in_b = in_ + b * total;
-    device SCALAR_T*       out_b = out + b * total;
+    // ulong base offsets: b * total overflows 32 bits on huge inputs.
+    device const SCALAR_T* in_b = in_ + (ulong)b * total;
+    device SCALAR_T*       out_b = out + (ulong)b * total;
 
     // Shift by the batch's first element before summing so the one-pass
     // variance doesn't lose precision to cancellation on large-DC inputs
@@ -112,8 +113,8 @@ kernel void apply_norm_gelu(
     float mean  = meanvar[b * 2 + 0];
     float scale = meanvar[b * 2 + 1];
 
-    device const SCALAR_T* x_b   = in_ + b * total_per_b;
-    device SCALAR_T*       out_b = out + b * total_per_b;
+    device const SCALAR_T* x_b   = in_ + (ulong)b * total_per_b;
+    device SCALAR_T*       out_b = out + (ulong)b * total_per_b;
 
     for (uint i = start + tid; i < end; i += tgs) {
         uint  c_idx = i / N;
